Fixed config not taking effect for float values
Fixes #28 most likely

+++ src/main/java/epicsquid/superiorshields/config/ShieldConfig.java
@@ -8,7 +8,7 @@ import java.util.Map;
 
 public class ShieldConfig implements IShieldConfig {
 
-	private final ForgeConfigSpec.ConfigValue<Float> CAPACITY;
+	private final ForgeConfigSpec.ConfigValue<Double> CAPACITY;
 	private final ForgeConfigSpec.ConfigValue<Integer> RATE;
 	private final ForgeConfigSpec.ConfigValue<Integer> DELAY;
 	private final IShieldType TYPE;
@@ -16,7 +16,7 @@ public class ShieldConfig implements IShieldConfig {
 	public ShieldConfig(ForgeConfigSpec.Builder builder, IShieldType type) {
 		builder.push(type.name().toLowerCase(Locale.ROOT) + "_shield");
 		TYPE = type;
-		CAPACITY = builder.comment("The amount of health the shield has. One point is half a shield icon (same as hearts.)").define("capacity", type.getDefaultCapacity());
+		CAPACITY = builder.comment("The amount of health the shield has. One point is half a shield icon (same as hearts.)").define("capacity", (double) type.getDefaultCapacity());
 		RATE = builder.comment("The amount of ticks between each attempt to recharge the shield bar.").define("recharge_rate", type.getDefaultRate());
 		DELAY = builder.comment("The amount of ticks without taking damage before the shield begins to recharge.").define("recharge_delay", type.getDefaultDelay());
 		builder.pop();
@@ -28,7 +28,7 @@ public class ShieldConfig implements IShieldConfig {
 
 	@Override
 	public float getCapacity() {
-		return CAPACITY.get();
+		return CAPACITY.get().floatValue();
 	}
 
 	@Override
